requirements.txt
aniso8601==9.0.1
anyio==3.6.1
api.ai==0.1.0
apiai==1.2.3
APScheduler==3.8.1
asgiref==3.5.2
attrs==21.2.0
backports.zoneinfo==0.2.1
bidict==0.21.2
cachelib==0.1.1
cachetools==3.1.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
charset-normalizer==2.0.12
click==7.1.1
colorama==0.4.5
colorclass==2.2.0
dialogflow==2.0.0
docopt==0.6.2
fastapi==0.78.0
fastapi-sessions==0.3.2
Flask==1.1.1
Flask-Cors==3.0.8
Flask-RESTful==0.3.9
flask-restplus==0.13.0
Flask-Session==0.3.1
Flask-SocketIO==4.1.1
Flask-SQLAlchemy==2.5.1
gevent==1.4.0
gevent-websocket==0.10.1
google-api-core==2.3.0
google-auth==2.3.3
google-cloud-dialogflow==2.10.0
googleapis-common-protos==1.54.0
greenlet==0.4.15
grpcio==1.42.0
grpcio-status==1.42.0
grpcio-tools==1.36.1
gunicorn==20.0.4
h11==0.13.0
httptools==0.4.0
idna==2.9
importlib-metadata==4.8.1
importlib-resources==5.4.0
itsdangerous==1.1.0
Jinja2==2.11.1
JsonForm==0.0.2
jsonschema==4.3.2
JsonSir==0.0.2
MarkupSafe==1.1.1
monotonic==1.5
mysql-connector-python==8.0.26
mysqlclient==2.0.3
numpy==1.18.2
packaging==20.3
pandas==1.1.5
pip-upgrader==1.4.15
proto-plus==1.19.8
protobuf==3.19.1
psutil==5.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.9.1
pyparsing==2.4.6
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.1
Python-EasyConfig==0.1.7
python-engineio==3.9.3
python-socketio==4.3.0
pytz==2019.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.23.0
Resource==0.2.1
rsa==4.0
schedule==1.1.0
six==1.14.0
sniffio==1.2.0
SQLAlchemy==1.4.25
starlette==0.19.1
terminaltables==3.1.0
typing-extensions==3.10.0.2
tzdata==2021.5
tzlocal==4.1
urllib3==1.25.8
uvicorn==0.17.6
watchgod==0.8.2
websockets==10.3
Werkzeug==1.0.0
zipp==3.6.0
zope.event==4.5.0
zope.interface==5.3.0
